Split lab context wearables
