+++ tsconfig.checkjs.json
@@ -113,6 +113,7 @@
     "js/import-review-draft.js",
     "js/import-review-row-actions.js",
     "js/lab-context.js",
+    "js/lab-context-wearables.js",
     "js/lens-actions.js",
     "js/lens-cache.js",
     "js/lens-library.js",
