+++ tsconfig.json
@@ -119,6 +119,7 @@
     "js/lab-entry.js",
     "js/lab-entry-mutations.js",
     "js/lab-context.js",
+    "js/lab-context-wearables.js",
     "js/lens-page-shell.js",
     "js/lens-pages.js",
     "js/lens.js",
